requirements.txt
arviz==0.12.0
cachetools==5.0.0
cftime==1.6.0
cycler==0.11.0
deprecat==2.1.1
dill==0.3.4
fastprogress==1.0.2
filelock==3.6.0
fonttools==4.31.2
kiwisolver==1.4.2
matplotlib==3.5.1
netCDF4==1.5.8
numpy==1.22.1
packaging==21.3
pandas==1.4.2
patsy==0.5.2
Pillow==9.1.0
polars==0.13.18
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
scipy==1.7.3
semver==2.13.0
six==1.16.0
Theano-PyMC==1.1.2
typing_extensions==4.1.1
wrapt==1.14.0
xarray==2022.3.0
